OnnRestApi/MainApi/serializers.py
from rest_framework import serializers
from MainApi.models import *


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ("id", "title", "price", "stock", "description")


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = ("id", "product", "address")
