requirements.txt
Django==3.2.4
djangorestframework==3.13.1
PyMySQL==1.0.2
